Fix UC_USELESS_CONDITION SpotBugs warnings and enforce rule in CI

+++ CodenameOne/src/com/codename1/ui/html/CSSParser.java
@@ -107,7 +107,7 @@ class CSSParser {
                 c = r.readCharFromReader();
             }
             c = r.readCharFromReader();
-            while (((byte) c) != -1 && isWhiteSpace(c)) { //skip white spaces
+            while (((int) c) != -1 && isWhiteSpace(c)) { //skip white spaces
                 c = r.readCharFromReader();
             }
         } else {
